// src/app/services/clients/device.client.ts
export interface DeviceEntity {
    id?: string | undefined;
    manufacturerID: number;
    model: string;
    ram: number;
    screenSize: number;
    operatingSystem: string;
}