// src/app/services/clients/manufacturer.client.ts
export interface Manufacturer {
    id: number;
    name: string;
}